Code Refactor
This was done in order to reduce the surface area I needed to modify when changing the data types over.

It's not related to transforming code for Burst/Job System, just done in order to make later passes slightly easier.

NB a couple of unused variables were trimmed in this commit too.