Map Class -> Struct
The Map class is shared between various classes in the sample. The data inside it will be needed by the FlowFieldGenerator and so we need to make sure it is a value type.

This simply moves all the data into a self contained struct, but leaves the Map as a class so it can still be shared.

Now, when we come to jobify the FlowFieldGenerator, the map data can be referenced by the job.