Jobified and Bursted Transform Code for BotManager
As Shown in the Presentation :

A struct is created to house the information needed by the transform task.
A temporary array (tempJob allocation) is made based on the number of bots
The data needed is copied into the array
The job is instantiated and Run() on the main thread
The results are copied back
The temporary allocation is freed.

Note - This differs from the live demonstration only in that the transform code was not placed into a static method